+++ ucte/ucte-network/src/test/java/com/powsybl/ucte/network/io/UcteRecordWriterTest.java
@@ -23,8 +23,8 @@ public class UcteRecordWriterTest {
         BufferedWriter bufferedWriter = new BufferedWriter(writer);
         UcteRecordWriter recordWriter = new UcteRecordWriter(bufferedWriter);
 
-        float input = 3.14f;
-        recordWriter.writeFloat(input, 0, 6);
+        double input = 3.14;
+        recordWriter.writeDouble(input, 0, 6);
         recordWriter.newLine();
         bufferedWriter.close();
 
@@ -37,8 +37,8 @@ public class UcteRecordWriterTest {
         BufferedWriter bufferedWriter = new BufferedWriter(writer);
         UcteRecordWriter recordWriter = new UcteRecordWriter(bufferedWriter);
 
-        float input = 3.1415927f;
-        recordWriter.writeFloat(input, 0, 6);
+        double input = 3.1415927;
+        recordWriter.writeDouble(input, 0, 6);
         recordWriter.newLine();
         bufferedWriter.close();
 
@@ -51,8 +51,8 @@ public class UcteRecordWriterTest {
         BufferedWriter bufferedWriter = new BufferedWriter(writer);
         UcteRecordWriter recordWriter = new UcteRecordWriter(bufferedWriter);
 
-        float input = -3.1f;
-        recordWriter.writeFloat(input, 0, 6);
+        double input = -3.1;
+        recordWriter.writeDouble(input, 0, 6);
         recordWriter.newLine();
         bufferedWriter.close();
 
@@ -65,8 +65,8 @@ public class UcteRecordWriterTest {
         BufferedWriter bufferedWriter = new BufferedWriter(writer);
         UcteRecordWriter recordWriter = new UcteRecordWriter(bufferedWriter);
 
-        float input = 0.0001f;
-        recordWriter.writeFloat(input, 0, 6);
+        double input = 0.0001;
+        recordWriter.writeDouble(input, 0, 6);
         recordWriter.newLine();
         bufferedWriter.close();
 
@@ -79,8 +79,8 @@ public class UcteRecordWriterTest {
         BufferedWriter bufferedWriter = new BufferedWriter(writer);
         UcteRecordWriter recordWriter = new UcteRecordWriter(bufferedWriter);
 
-        float input = Float.NaN;
-        recordWriter.writeFloat(input, 0, 6);
+        double input = Double.NaN;
+        recordWriter.writeDouble(input, 0, 6);
         recordWriter.newLine();
         bufferedWriter.close();
 
@@ -93,8 +93,8 @@ public class UcteRecordWriterTest {
         BufferedWriter bufferedWriter = new BufferedWriter(writer);
         UcteRecordWriter recordWriter = new UcteRecordWriter(bufferedWriter);
 
-        float input = 12345678f;
-        recordWriter.writeFloat(input, 0, 6);
+        double input = 12345678;
+        recordWriter.writeDouble(input, 0, 6);
         recordWriter.newLine();
         bufferedWriter.close();
     }
@@ -105,8 +105,8 @@ public class UcteRecordWriterTest {
         BufferedWriter bufferedWriter = new BufferedWriter(writer);
         UcteRecordWriter recordWriter = new UcteRecordWriter(bufferedWriter);
 
-        float input = -12345678f;
-        recordWriter.writeFloat(input, 0, 6);
+        double input = -12345678;
+        recordWriter.writeDouble(input, 0, 6);
         recordWriter.newLine();
         bufferedWriter.close();
     }
@@ -117,8 +117,8 @@ public class UcteRecordWriterTest {
         BufferedWriter bufferedWriter = new BufferedWriter(writer);
         UcteRecordWriter recordWriter = new UcteRecordWriter(bufferedWriter);
 
-        float input = 999999.9f;
-        recordWriter.writeFloat(input, 0, 6);
+        double input = 999999.9;
+        recordWriter.writeDouble(input, 0, 6);
         recordWriter.newLine();
         bufferedWriter.close();
 
@@ -131,8 +131,8 @@ public class UcteRecordWriterTest {
         BufferedWriter bufferedWriter = new BufferedWriter(writer);
         UcteRecordWriter recordWriter = new UcteRecordWriter(bufferedWriter);
 
-        float input = 1000000.1f;
-        recordWriter.writeFloat(input, 0, 6);
+        double input = 1000000.1;
+        recordWriter.writeDouble(input, 0, 6);
         recordWriter.newLine();
         bufferedWriter.close();
     }
@@ -143,8 +143,8 @@ public class UcteRecordWriterTest {
         BufferedWriter bufferedWriter = new BufferedWriter(writer);
         UcteRecordWriter recordWriter = new UcteRecordWriter(bufferedWriter);
 
-        float input = -99999.9f;
-        recordWriter.writeFloat(input, 0, 6);
+        double input = -99999.9;
+        recordWriter.writeDouble(input, 0, 6);
         recordWriter.newLine();
         bufferedWriter.close();
 
@@ -157,8 +157,8 @@ public class UcteRecordWriterTest {
         BufferedWriter bufferedWriter = new BufferedWriter(writer);
         UcteRecordWriter recordWriter = new UcteRecordWriter(bufferedWriter);
 
-        float input = -100000.1f;
-        recordWriter.writeFloat(input, 0, 6);
+        double input = -100000.1;
+        recordWriter.writeDouble(input, 0, 6);
         recordWriter.newLine();
         bufferedWriter.close();
     }
